訂正版#2

+++ src/apple.c
@@ -24,7 +24,7 @@ int main(){
  
 
     
-    int Sum=0;
+    long long int Sum=0;
     for(i=0; i<n; i++){
         Sum=Sum+A[i];
     }
